test: skip archive source if token missing
As explained in https://github.com/dependabot/dependabot-core/issues/3253#issuecomment-797125425,
dependabot PRs no longer have access to secrets. This means that the
corresponding env var will be defined, but empty. Skip the test on the
assumption that we have enough coverage elsewhere to catch breakage
in requests and its deps